Move tests to test_pipeline
