+++ qiime2/sdk/tests/test_pipeline.py
@@ -214,6 +214,86 @@ class TestPipeline(unittest.TestCase):
             self.assertEqual(single_int.type, SingleInt)
             self.assertEqual(single_int.view(int), 4)
 
+    def test_de_facto_list_arg(self):
+        pipeline = self.plugin.pipelines['de_facto_list_pipeline']
+
+        exp = {'0': 0, '1': 1, '2': 2}
+
+        ret = pipeline()
+        obs = qiime2.sdk.util.view_collection(ret.output, int)
+
+        self.assertEqual(obs, exp)
+
+    def test_de_facto_list_arg_parallel(self):
+        pipeline = self.plugin.pipelines['de_facto_list_pipeline']
+
+        exp = {'0': 0, '1': 1, '2': 2}
+
+        ret = pipeline.parallel()._result()
+        obs = qiime2.sdk.util.view_collection(ret.output, int)
+
+        self.assertEqual(obs, exp)
+
+    def test_de_facto_list_kwarg(self):
+        pipeline = self.plugin.pipelines['de_facto_list_pipeline']
+
+        exp = {'0': 0, '1': 1, '2': 2}
+
+        ret = pipeline(kwarg=True)
+        obs = qiime2.sdk.util.view_collection(ret.output, int)
+
+        self.assertEqual(obs, exp)
+
+    def test_de_facto_list_kwarg_parallel(self):
+        pipeline = self.plugin.pipelines['de_facto_list_pipeline']
+
+        exp = {'0': 0, '1': 1, '2': 2}
+
+        ret = pipeline.parallel(kwarg=True)._result()
+        obs = qiime2.sdk.util.view_collection(ret.output, int)
+
+        self.assertEqual(obs, exp)
+
+    def test_de_facto_dict_arg(self):
+        pipeline = self.plugin.pipelines['de_facto_dict_pipeline']
+
+        exp = {'1': 0, '2': 1, '3': 2}
+
+        ret = pipeline()
+        obs = qiime2.sdk.util.view_collection(ret.output, int)
+
+        self.assertEqual(obs, exp)
+
+    def test_de_facto_dict_arg_parallel(self):
+        pipeline = self.plugin.pipelines['de_facto_dict_pipeline']
+
+        exp = {'1': 0, '2': 1, '3': 2}
+
+        ret = pipeline.parallel()._result()
+        obs = qiime2.sdk.util.view_collection(ret.output, int)
+
+        self.assertEqual(obs, exp)
+
+    def test_de_facto_dict_kwarg(self):
+        pipeline = self.plugin.pipelines['de_facto_dict_pipeline']
+
+        exp = {'1': 0, '2': 1, '3': 2}
+
+        ret = pipeline(kwarg=True)
+        obs = qiime2.sdk.util.view_collection(ret.output, int)
+
+        self.assertEqual(obs, exp)
+
+    def test_de_facto_dict_kwarg_parallel(self):
+        pipeline = self.plugin.pipelines['de_facto_dict_pipeline']
+
+        exp = {'1': 0, '2': 1, '3': 2}
+
+        ret = pipeline.parallel(kwarg=True)._result()
+        obs = qiime2.sdk.util.view_collection(ret.output, int)
+
+        self.assertEqual(obs, exp)
+
     def test_failing_from_arity(self):
         for call in self.iter_callables('failing_pipeline'):
             with self.assertRaisesRegex(TypeError, 'match number.*3.*1'):
@@ -248,6 +328,34 @@ class TestPipeline(unittest.TestCase):
         for call in self.iter_callables('failing_pipeline'):
             with self.assertRaisesRegex(ValueError, r'action.*not\%a\$method'):
                 call(self.int_sequence, break_from='no-action')
+
+    def test_fail_de_facto_list_arg_mixed(self):
+        pipeline = self.plugin.pipelines['de_facto_list_pipeline']
+
+        with self.assertRaisesRegex(
+                ValueError, 'Collection has mixed proxies and artifacts.*'):
+            pipeline.parallel(non_proxies=True)._result()
+
+    def test_fail_de_facto_list_kwarg_mixed(self):
+        pipeline = self.plugin.pipelines['de_facto_list_pipeline']
+
+        with self.assertRaisesRegex(
+                ValueError, 'Collection has mixed proxies and artifacts.*'):
+            pipeline.parallel(kwarg=True, non_proxies=True)._result()
+
+    def test_fail_de_facto_dict_arg_mixed(self):
+        pipeline = self.plugin.pipelines['de_facto_dict_pipeline']
+
+        with self.assertRaisesRegex(
+                ValueError, 'Collection has mixed proxies and artifacts.*'):
+            pipeline.parallel(non_proxies=True)._result()
+
+    def test_fail_de_facto_dict_kwarg_mixed(self):
+        pipeline = self.plugin.pipelines['de_facto_dict_pipeline']
+
+        with self.assertRaisesRegex(
+                ValueError, 'Collection has mixed proxies and artifacts.*'):
+            pipeline.parallel(kwarg=True, non_proxies=True)._result()
 
 
 if __name__ == '__main__':
